Looks like an integer is actually a long, wow

+++ interaction-declarations/src/main/kotlin/net/perfectdreams/discordinteraktions/declarations/commands/slash/options/CommandOptions.kt
@@ -23,13 +23,13 @@ open class CommandOptions {
         description
     )
 
-    fun integer(name: String, description: String) = argument<Int>(
+    fun integer(name: String, description: String) = argument<Long>(
         CommandOptionType.Integer,
         name,
         description
     )
 
-    fun optionalInteger(name: String, description: String) = argument<Int?>(
+    fun optionalInteger(name: String, description: String) = argument<Long?>(
         CommandOptionType.NullableInteger,
         name,
         description
